Fix the bug
